Move logic to vuex store
