+++ src/vuex/modules/pushNotifications.js
@@ -6,13 +6,9 @@ let messaging
 const initializeFirebase = () => {
   firebase.initializeApp({
     apiKey: "AIzaSyCrA4mq9v926h3aX9mfkLlzUSRbjFude14",
-    authDomain: "lunie-push-notifications.firebaseapp.com",
-    databaseURL: "https://lunie-push-notifications.firebaseio.com",
     projectId: "lunie-push-notifications",
-    storageBucket: "lunie-push-notifications.appspot.com",
     messagingSenderId: "783884833065",
-    appId: "1:783884833065:web:ea02768959989b9218a738",
-    measurementId: "G-S3JEYPBQWB"
+    appId: "1:783884833065:web:ea02768959989b9218a738"
   })
 
   messaging = firebase.messaging()
@@ -58,35 +54,8 @@ const registerDevice = async (token, activeNetworks) => {
   }
 }
 
-/**
- * Retrieve active networks from localstorage via session keys
- *
- * @param {[ Object ]} networks
- * @param network.id network identifier such as "cosmos-hub-testnet"
- */
-const getActiveNetworks = networks => {
-  let activeNetworks = []
-  networks.forEach(network => {
-    // Session object: { address: string, sessionType: string (e.g. ledger)}
-    const networkObject = JSON.parse(
-      localStorage.getItem(`session_${network.id}`)
-    )
-
-    // Only store network object if it has an associated address
-    if (networkObject) {
-      activeNetworks.push({
-        address: networkObject.address,
-        networkId: network.id
-      })
-    }
-  })
-
-  return activeNetworks
-}
-
 export default {
   initializeFirebase,
   askPermissionAndRegister,
-  registerDevice,
-  getActiveNetworks
+  registerDevice
 }
